Dashboard routes and merge with master.
