--- src/pages/Dashboard/styled.ts
@@ -1,14 +1,0 @@
-import styled from "styled-components";
-
-const ContainerDashboard = styled.div`
-  width: 100%;
-  height: 100%;
-  background-color: #27364b;
-  color: white;
-  background-image: linear-gradient(175deg, #27364b 5%, #0e1318 100%);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-export { ContainerDashboard };
