+++ src/pages/DashboardPage/style.ts
@@ -1,14 +1,14 @@
 import styled from "styled-components";
 
 export const DashboardPageContainer = styled.div`
+  width: 100%;
+  height: 100%;
+  background-color: #27364b;
+  color: white;
+  background-image: linear-gradient(175deg, #27364b 5%, #0e1318 100%);
   display: flex;
   justify-content: center;
   align-items: center;
-  background-color: #27364b;
-  flex-direction: row;
-  width: 100%;
-  height: 100%;
-  background-image: linear-gradient(175deg, #27364b 5%, #0e1318 100%);
 `;
 
 export const DashboardPageSideBarContainer = styled.div`
